requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aioredis==2.0.1
aiosignal==1.3.1
aiosqlite==0.20.0
alembic==1.13.2
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.12.3
boto3==1.42.39
botocore==1.42.39
certifi==2026.1.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
deepdiff==7.0.1
distro==1.9.0
dnspython==2.6.1
email-validator==2.2.0
emoji==2.12.1
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.112.0
fastapi-cli==0.0.5
ffmpy==0.4.0
filelock==3.15.4
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.0.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.1
httpx==0.28.1
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.11
importlib-resources==6.4.3
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
jinja2==3.1.4
jiter==0.12.0
jmespath==1.1.0
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jupyter-client==8.6.2
jupyter-core==5.7.2
kiwisolver==1.4.5
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-openai==0.1.21
langchain-redis==0.0.4
langchain-text-splitters==0.2.2
langdetect==1.0.9
langgraph==0.2.3
langgraph-checkpoint==1.0.2
langgraph-checkpoint-sqlite==1.0.0
langsmith==0.1.147
loguru==0.7.2
lxml==5.2.2
mako==1.3.5
markdown==3.6
markdown-it-py==3.0.0
markupsafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mypy==1.11.1
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
onnxruntime==1.18.1
openai==1.109.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
ordered-set==4.1.0
orjson==3.11.6
packaging==24.2
pandas==2.2.2
parso==0.8.4
passlib==1.7.4
pdf2image==1.17.0
pdfminer-six==20231228
pdfplumber==0.11.3
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prompt-toolkit==3.0.47
protobuf==5.27.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.4.0
pydub==0.25.1
pygments==2.18.0
pyjwt==2.9.0
pymupdf==1.24.9
pymupdfb==1.24.9
pyparsing==3.1.2
pypdf==4.3.1
pypdfium2==4.30.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
python-pptx==1.0.2
python-ulid==2.7.0
pytz==2024.1
pyyaml==6.0.3
pyzmq==26.1.0
rapidfuzz==3.9.6
rapidocr-onnxruntime==1.3.24
redis==5.1.0
redisvl==0.3.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rich==13.7.1
ruff==0.5.7
s3transfer==0.16.0
scalar-fastapi==1.3.0
semantic-version==2.10.0
setuptools==73.0.1
shapely==2.0.5
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.5
sqlalchemy==2.0.32
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.37.2
sympy==1.13.1
tabulate==0.9.0
tavily-python==0.4.0
tenacity==8.5.0
tiktoken==0.12.0
tomlkit==0.12.0
tornado==6.4.1
tqdm==4.67.2
traitlets==5.14.3
typer==0.12.3
types-aiofiles==24.1.0.20240626
types-beautifulsoup4==4.12.0.20240511
types-html5lib==1.1.11.20240806
types-pyyaml==6.0.12.20240808
types-requests==2.32.0.20240712
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2024.1
unstructured==0.14.9
unstructured-client==0.25.4
urllib3==2.6.3
uvicorn==0.30.5
uvloop==0.19.0
watchfiles==0.23.0
wcwidth==0.2.13
websockets==12.0
wrapt==1.16.0
xlrd==2.0.1
xlsxwriter==3.2.0
yarl==1.9.4
zstandard==0.23.0
